chore: release version 1.0.0-beta.13
