+++ package.json
@@ -24,7 +24,8 @@
     "graphql:codegen": "graphql-codegen --config codegen.yml",
     "lint": "eslint . --ext .ts,.vue --fix",
     "prepare": "husky install",
-    "publish:packages": "node ./scripts/publish.js",
+    "publish:api-client": "node ./scripts/publishApi.js",
+    "publish:composables": "node ./scripts/publishComposable.js",
     "start": "cd packages/theme && yarn start",
     "test": "yarn test:api-client && yarn test:composables && yarn test:theme",
     "test:api-client": "cd packages/api-client && yarn test --passWithNoTests",
