+++ scripts/publishApi.js
@@ -1,10 +1,8 @@
-/* eslint-disable unicorn/no-process-exit */
-const { publishPackages } = require('./lib/publishPackages');
+/* eslint-disable unicorn/no-process-exit, unicorn/prefer-module */
+const { publishPackages } = require('./lib/publishApi');
 
 const myArgs = process.argv.slice(2);
 const labels = myArgs[0].split('|');
-
-console.log(publishPackages);
 
 publishPackages(labels)
   .then(console.log)
